requirements.txt
selenium~=3.141.0
Appium-Python-Client~=1.1.0